+++ backend/warranty/src/main/java/edu/uth/warranty/model/WarrantyClaim.java
@@ -1,13 +1,21 @@
 package edu.uth.warranty.model;
 
-import jakarta.persistence.*;
-import java.time.LocalDateTime;
 import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.NoArgsConstructor;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 @Entity
 @Getter
 @Setter
@@ -58,11 +66,11 @@ public class WarrantyClaim {
     private String approvalStatus;
 
     @Column(name = "current_mileage")
-    private Integer currentMileage;
+    private Long currentMileage;
 
     public WarrantyClaim(Long claimId) {
         this.claimId = claimId;
     }
 
 
-}   
+}
